BC-72 | Started on splitting out SubmissionValidationService into multiple modules following single-responsibility principle
